requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.18.0
altair==5.0.1
annotated-types==0.6.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
Babel==2.13.0
backcall==0.2.0
bce-python-sdk==0.8.92
beautifulsoup4==4.12.2
behave==1.2.6
bleach==4.1.0
blinker==1.6.3
boltons==23.0.0
Bottleneck==1.3.5
Brotli==1.1.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
conda==23.9.0
conda-content-trust==0.1.3
conda-libmamba-solver==23.9.2
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
contourpy==1.0.5
cryptography==39.0.1
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.0
cytoolz==0.12.0
dask==2023.6.0
dataclasses==0.6
datasets==2.14.5
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
distlib==0.3.7
distro==1.8.0
djitellopy==2.5.0
easydict==1.10
entrypoints==0.4
exceptiongroup==1.1.3
execnet==2.0.2
executing==0.8.3
fastapi==0.104.0
fasteners==0.19
fastjsonschema==2.16.2
ffmpy==0.3.1
filelock==3.12.4
flake8==6.1.0
Flask==3.0.0
flask-babel==4.0.0
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.37
google-api-core==2.14.0
google-api-python-client==2.107.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==0.5.2
google-pasta==0.2.0
googleapis-common-protos==1.61.0
gotrue==1.2.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.48.2
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.18.0
hyperframe==6.0.1
idna==3.4
imageai==3.0.3
imagecodecs==2023.1.23
imageio==2.31.4
imgaug==0.4.0
importlib-metadata==6.0.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-server==1.23.4
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.5
kaitaistruct==0.10
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libmambapy==1.5.1
locket==1.0.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.0
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.12.2
munkres==1.1.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
netaddr==0.9.0
netifaces==0.11.0
networkx==3.2.1
nlu==5.1.0
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.14.1
openai==1.2.3
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.9
outcome==1.2.0
packaging==23.2
paddle-bfloat==0.1.7
paddle2onnx==1.1.0
paddlefsl==1.1.0
paddlehub==2.4.0
paddlenlp==2.6.1
paddlepaddle==2.5.1
pandas==2.1.1
pandocfilters==1.5.0
parameterized==0.9.0
parse==1.19.1
parse-type==0.6.2
parso==0.8.3
partd==1.4.1
pdbp==1.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3
platformdirs==3.10.0
pluggy==1.0.0
ply==3.11
postgrest==0.11.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
py4j==0.10.9
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycodestyle==2.11.1
pycosat==0.6.6
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.7.1
pydub==0.25.1
pyflakes==3.1.0
pygame==2.5.2
Pygments==2.15.1
PyJWT==2.4.0
Pympler==0.9
pynose==1.4.8
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
pyspark==3.0.2
pytest==7.4.2
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==23.2.0
qtconsole==5.4.2
QtPy==2.2.0
rarfile==4.1
realtime==1.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.10.6
rsa==4.7.2
ruamel.yaml==0.17.21
safetensors==0.4.0
sbvirtualdisplay==1.2.0
scikit-image==0.20.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
selenium==4.13.0
selenium-authenticated-proxy==1.0.1
selenium-base==4.19.2
selenium-wire==5.1.0
seleniumbase==4.19.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
seqeval==1.2.2
setuptools==68.0.0
shapely==2.0.1
sip==6.7.12
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.5
spark-nlp==5.1.4
stack-data==0.2.0
starlette==0.27.0
storage3==0.6.1
streamlit==1.24.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
sympy==1.12
tabcompleter==1.3.0
tello-python==1.1.6
tellopy==0.6.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
termcolor==2.1.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.4.12
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
trio==0.22.2
trio-websocket==0.11.1
triton==2.1.0
truststore==0.8.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==2.1
ultralytics==8.0.203
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.23.2
validators==0.18.2
virtualenv==20.24.5
visualdl==2.5.3
watchdog==2.1.6
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==0.58.0
websockets==10.4
Werkzeug==2.2.3
wheel==0.41.2
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
xxhash==3.4.1
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.11.0
zstandard==0.19.0
pantalaimon[ui]
